add buttonsContainer
