+++ src/scenes/projects/buttonsContainer.module.scss
@@ -1,0 +1,7 @@
+.buttonsContainer {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-around;
+    padding: 16px 0;
+    cursor: auto;
+  }
