+++ src/scenes/projects/cardContent.module.scss
@@ -5,7 +5,7 @@
   flex-direction: row;
   flex-wrap: nowrap;
     height: auto; //TODO change when content ready
-    width: calc(300vw - 24px); //40px for margins
+    width: calc(300vw - 8px); //40px for margins
 }
 
 .backgroundPastabook {
