+++ src/scenes/projects/slide3.module.scss
@@ -4,16 +4,30 @@
   display: flex;
   flex-direction: column;
   justify-content: space-around;
+  align-items: flex-end;
   width: 100vw;
   perspective: 2000px;
 
   .screenDesktop {
     width: 80vw;
     height: auto;
-    margin: 48px 4px 16px 16px;
-    transform: rotateY(10deg) rotateX(5deg);
+    margin: 48px 16px 16px 4px;
+    transform: rotateY(-10deg) rotateX(-5deg);
     filter: brightness(0.9) contrast(1.1);
     box-shadow: 10px 10px 20px rgba(0, 0, 0, 0.5);
     box-shadow: 5px 5px 15px rgba(0, 0, 0, 0.4);
+  }
+
+  .positionSetter {
+    width: 90vw;
+  }
+}
+
+
+@media screen and (min-width: $tablet) {
+  .container{
+    .screenDesktop{
+      margin: 48px 32px 16px 4px;
+    }
   }
 }
